Fix: updated javatraverser2

+++ javatraverser2/java/mds/data/descriptor_s/Pointer.java
@@ -16,7 +16,7 @@ public final class Pointer extends Descriptor_S<Number> implements DATA<Number>,
     }
 
     public static final Pointer NULL() {
-        return new Pointer(ByteBuffer.wrap(new byte[]{8, 0, DTYPE.POINTER, Descriptor_S.CLASS, 8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}).order(Descriptor.BYTEORDER));
+        return new Pointer(ByteBuffer.wrap(new byte[]{8, 0, DTYPE.POINTER.toByte(), Descriptor_S.CLASS, 8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}).order(Descriptor.BYTEORDER));
     }
 
     public Pointer(final ByteBuffer b){
